backend/pyproject.toml
[project]
name = "mental-health-backend"
version = "3.0.0"
description = "Mental Health AI Backend - Day 3 Database Enhanced"
requires-python = ">=3.11"
authors = [
    {name = "Enthusiast-AD", email = "enthusiast@example.com"}
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "psycopg2-binary>=2.9.9",
    "bcrypt>=4.1.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "pydantic[email]>=2.5.0,<3.0.0",
    "websockets>=12.0",
    "httpx>=0.25.2",
    "nltk>=3.8.1",
    "textblob>=0.17.1",
    "pytest>=7.4.3",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "black>=23.0.0",
    "isort>=5.12.0"
]
